added tests for custom method in entity policy
